moved base hub class and has ui interface into a folder
