+++ SignalRWebPack/src/js/Hubs/Base/IHasUIHub.ts
