+++ SignalRWebPack/src/js/Hubs/Base/baseHub.ts
