+++ SignalRWebPack/src/js/tsconfig.json
@@ -14,8 +14,8 @@
     "files": [
         "index.ts",
         "Hubs/initHubs.ts",
-        "Hubs/baseHub.ts",
-        "Hubs/IHasUIHub.ts",
+        "Hubs/Base/baseHub.ts",
+        "Hubs/Base/IHasUIHub.ts",
         "Hubs/ControlHubs/aslControlServerHub.ts",
         "Hubs/DataHubs/libraryDataServerHub.ts",
         "Hubs/DataHubs/vikDataServerHub.ts"
